# packages/openhands-agent-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenHands Agent Server",
    "version": "0.1.0",
    "description": "OpenHands Agent Server - REST/WebSocket interface for OpenHands AI Agent"
  },
  "paths": {
...
    "/api/conversations/search": {
      "get": {
        "tags": [
          "Conversations"
        ],
        "summary": "Search conversations",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of items to return.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "page_id",
            "in": "query",
            "required": false,
            "description": "Opaque pagination cursor.",
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "workspace": {
                            "type": "object",
                            "properties": {
                              "kind": {
                                "default": "LocalWorkspace",
                                "type": "string"
                              },
                              "working_dir": {
                                "default": "workspace/project",
                                "type": "string"
                              }
                            },
                            "required": [
                              "kind",
                              "working_dir"
                            ],
                            "additionalProperties": {}
                          },
                          "persistence_dir": {
                            "default": "workspace/conversations",
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "max_iterations": {
                            "default": 500,
                            "type": "integer",
                            "exclusiveMinimum": 0,
                            "maximum": 9007199254740991
                          },
                          "stuck_detection": {
                            "default": true,
                            "type": "boolean"
                          },
                          "execution_status": {
                            "type": "string",
                            "enum": [
                              "idle",
                              "running",
                              "paused",
                              "finished",
                              "error",
                              "stuck",
                              "deleting"
                            ]
                          },
                          "confirmation_policy": {},
                          "security_analyzer": {
                            "anyOf": [
                              {},
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "activated_knowledge_skills": {
                            "default": [],
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "invoked_skills": {
                            "default": [],
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "blocked_actions": {
                            "default": {},
                            "type": "object",
                            "propertyNames": {
                              "type": "string"
                            },
                            "additionalProperties": {
                              "type": "string"
                            }
                          },
                          "blocked_messages": {
                            "default": {},
                            "type": "object",
                            "propertyNames": {
                              "type": "string"
                            },
                            "additionalProperties": {
                              "type": "string"
                            }
                          },
                          "last_user_message_id": {
                            "default": null,
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "stats": {
                            "default": {},
                            "type": "object",
                            "propertyNames": {
                              "type": "string"
                            },
                            "additionalProperties": {}
                          },
                          "secret_registry": {
                            "default": {},
                            "type": "object",
                            "propertyNames": {
                              "type": "string"
                            },
                            "additionalProperties": {}
                          },
                          "agent_state": {
                            "default": {},
                            "type": "object",
                            "propertyNames": {
                              "type": "string"
                            },
                            "additionalProperties": {}
                          },
                          "hook_config": {
                            "default": null,
                            "anyOf": [
                              {},
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "title": {
                            "default": null,
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "metrics": {
                            "default": null,
                            "anyOf": [
                              {},
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "created_at": {
                            "type": "string"
                          },
                          "updated_at": {
                            "type": "string"
                          },
                          "tags": {
                            "default": {},
                            "type": "object",
                            "propertyNames": {
                              "type": "string"
                            },
                            "additionalProperties": {
                              "type": "string"
                            }
                          },
                          "current_model_id": {
                            "default": null,
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "available_models": {
                            "default": [],
                            "type": "array",
                            "items": {}
                          },
                          "supports_runtime_model_switch": {
                            "default": false,
                            "type": "boolean"
                          },
                          "launched_agent_profile": {
                            "default": null,
                            "anyOf": [
                              {},
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "agent": {},
                          "client_tools": {
                            "default": [],
                            "type": "array",
                            "items": {}
                          }
                        },
                        "required": [
                          "id",
                          "workspace",
                          "persistence_dir",
                          "max_iterations",
                          "stuck_detection",
                          "execution_status",
                          "activated_knowledge_skills",
                          "invoked_skills",
                          "blocked_actions",
                          "blocked_messages",
                          "last_user_message_id",
                          "stats",
                          "secret_registry",
                          "agent_state",
                          "hook_config",
                          "title",
                          "metrics",
                          "created_at",
                          "updated_at",
                          "tags",
                          "current_model_id",
                          "available_models",
                          "supports_runtime_model_switch",
                          "launched_agent_profile",
                          "client_tools"
                        ],
                        "additionalProperties": {}
                      }
                    },
                    "next_page_id": {
                      "default": null,
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "items",
                    "next_page_id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/conversations/{conversation_id}/agent_final_response": {
      "get": {
        "tags": [
          "Conversations"
        ],
        "summary": "Get final agent response",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "response": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "response"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Error"
          }
        }
      }
    },
    "/api/conversations/{conversation_id}/pause": {
      "post": {
        "tags": [
          "Conversations"
        ],
        "summary": "Pause conversation",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "success": {
                      "default": true,
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Error"
          }
        }
      }
    },
    "/api/conversations/{conversation_id}/interrupt": {
      "post": {
        "tags": [
          "Conversations"
        ],
        "summary": "Interrupt conversation",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "success": {
                      "default": true,
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Error"
          }
        }
      }
    },
    "/api/conversations/{conversation_id}/run": {
      "post": {
        "tags": [
          "Conversations"
        ],
        "summary": "Run conversation",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "success": {
                      "default": true,
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Error"
          },
          "409": {
            "description": "Error"
          }
        }
      }
    },
    "/api/conversations/{conversation_id}/goal": {
      "post": {
        "tags": [
          "Conversations"
        ],
        "summary": "Start goal loop",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "objective": {
                    "type": "string"
                  },
                  "max_iterations": {
                    "default": 10,
                    "type": "integer",
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991
                  }
                },
                "required": [
                  "objective",
                  "max_iterations"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "success": {
                      "default": true,
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Error"
          },
          "501": {
            "description": "Error"
          }
        }
      }
    },
    "/api/conversations/{conversation_id}/goal/stop": {
      "post": {
        "tags": [
          "Conversations"
        ],
        "summary": "Stop goal loop",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "success": {
                      "default": true,
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Error"
          },
          "501": {
            "description": "Error"
          }
        }
      }
    },
    "/api/conversations/{conversation_id}/goal/resume": {
      "post": {
        "tags": [
          "Conversations"
        ],
        "summary": "Resume goal loop",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "success": {
                      "default": true,
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Error"
          },
          "501": {
            "description": "Error"
          }
        }
      }
    },
    "/api/conversations/{conversation_id}/secrets": {
      "post": {
        "tags": [
          "Conversations"
        ],
        "summary": "Update keychain-backed conversation secrets",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "secrets": {
                    "type": "object",
                    "propertyNames": {
                      "type": "string"
                    },
                    "additionalProperties": {}
                  }
                },
                "required": [
                  "secrets"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "success": {
                      "default": true,
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Error"
          },
          "404": {
            "description": "Error"
          }
        }
      }
    },
    "/api/conversations/{conversation_id}/confirmation_policy": {
      "post": {
        "tags": [
          "Conversations"
        ],
        "summary": "Accepted deviation: confirmation policy is intentionally unsupported",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "policy": {}
                },
                "required": [
                  "policy"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "404": {
            "description": "Error"
          },
          "410": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "accepted_deviation": {
                      "type": "boolean",
                      "const": true
                    },
                    "feature": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "detail",
                    "accepted_deviation",
                    "feature"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/conversations/{conversation_id}/ask_agent": {
      "post": {
        "tags": [
          "Conversations"
        ],
        "summary": "Ask agent out of band",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "question": {
                    "type": "string"
                  }
                },
                "required": [
                  "question"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "response": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "response"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Error"
          },
          "501": {
            "description": "Error"
          }
        }
      }
    },
...
  }
}